add dockerfiles and buildind script

+++ src/main/java/de/app/subh/dataAccessObjects/DBWriter.java
@@ -16,7 +16,7 @@ import de.app.subh.utilities.EnryptionHelper;
 /**
  * class used for writing ( adding and deleting ) entities in the data base
  * 
- * @author lucien 
+ * @author lucien and @author francine
  *
  */
 
@@ -27,7 +27,8 @@ public class DBWriter {
 	private DBReader dbReader = new DBReader();
 	private EntityTransaction transaction = dbReader.getEntityManager().getTransaction();;
 
-	public DBWriter() {}
+	public DBWriter() {
+	}
 
 	/**
 	 * add a new user in the data base
@@ -51,27 +52,29 @@ public class DBWriter {
 	 */
 	public void deleteUser(User user) {
 
-		transaction.begin();
-		//dbReader.getEntityManager().merge(user);
-		dbReader.getEntityManager().remove(user);
-		transaction.commit();
+		if (Objects.nonNull(user)) {
+			transaction.begin();
+			dbReader.getEntityManager().remove(user);
+			transaction.commit();
+		} else
+			System.out.println("the user is null");
 	}
-	
+
 	/**
 	 * Add a new book in the data base
+	 * 
 	 * @param book
 	 */
 	public void addBook(Book book) {
-		
-		if(Objects.nonNull(book)) {
+
+		if (Objects.nonNull(book)) {
 			transaction.begin();
 			dbReader.getEntityManager().persist(book);
 			transaction.commit();
-		}
-		else
+		} else
 			System.out.println("the book is null");
 	}
-	
+
 	/**
 	 * delete a book in the data base
 	 * 
@@ -83,17 +86,19 @@ public class DBWriter {
 		dbReader.getEntityManager().remove(book);
 		transaction.commit();
 	}
-	
+
 	/**
 	 * create a new user and examine if his data are right
+	 * 
 	 * @param toCreate
 	 * @param unhashedPassword
 	 * @throws UserEmailException
 	 * @throws UserPasswordException
 	 * @throws EntityExistsException
 	 */
-	
-	public void createUser(User toCreate, String unhashedPassword) throws UserEmailException, UserPasswordException, EntityExistsException {
+
+	public void createUser(User toCreate, String unhashedPassword)
+			throws UserEmailException, UserPasswordException, EntityExistsException {
 		String email = toCreate.getEmail();
 		if (email == null || email.trim().length() == 0) {
 			throw new UserEmailException("E-Mail muss angegeben werden!");
@@ -105,61 +110,62 @@ public class DBWriter {
 		if (dbReader.findUserByNameOrEmail(toCreate.getUsername(), email) != null) {
 			throw new EntityExistsException("Benutzer existiert bereits!");
 		}
-		
+
 		toCreate.setSalt(EnryptionHelper.generateSalt());
 		toCreate.setPassword(EnryptionHelper.hashPwAndSalt(unhashedPassword, toCreate.getSalt()));
 
-		//persist the new user in the data base
+		// persist the new user in the data base
 		addUser(toCreate);
 	}
-	
+
 	/**
 	 * methode used when a user borrows one book
 	 */
 	public void borrowBook(Book bookMark, User userMark) {
-		
+
 		bookMark.setStatus("Borrowed");
 		bookMark.setBorrower(userMark.getUsername());
 		updateBook(bookMark);
-		
+
 	}
-	
+
 	public void giveBackBook(Book bookMark) {
-		
+
 		bookMark.setStatus("Available");
 		bookMark.setBorrower("");
 		updateBook(bookMark);
-		
+
 	}
-	
+
 	/**
 	 * update a a book given as parameter
+	 * 
 	 * @param book
 	 */
 	public void updateBook(final Book book) {
-		if(Objects.nonNull(book)) {
+		if (Objects.nonNull(book)) {
 			transaction.begin();
 			dbReader.getEntityManager().merge(book);
 //			dbReader.getEntityManager().persist(book);
 			transaction.commit();
-		}else
+		} else
 			System.out.println("the book is null !");
 	}
-	
+
 	/**
 	 * update a user given as parameter
+	 * 
 	 * @param user
 	 */
 	public void updateUser(final User user) {
-		if(Objects.nonNull(user)) {
+		if (Objects.nonNull(user)) {
 			transaction.begin();
 			dbReader.getEntityManager().merge(user);
 			transaction.commit();
-		}else
+		} else
 			System.out.println("the user is null !");
 	}
-	
-	
+
 	// Setters and Getters
 
 	public DBReader getDbReader() {
